protocols/dice: tcat: code cleanup for stream format section operation
This commit adds alternative functions to read and write parameters in
stream format section of TCAT protocol extension.
